+++ tests/dwd/observations/test_api_data.py
@@ -1,5 +1,6 @@
 from io import StringIO
 from pathlib import Path
+from datetime import datetime
 
 import pytest
 import pandas as pd
@@ -156,25 +157,6 @@ def test_dwd_observation_data_time_input():
         end_date=pd.Timestamp("1971-01-01"),
     )
 
-    request = DWDObservationData(
-        station_ids=[1],
-        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
-        resolution=DWDObservationResolution.DAILY,
-        periods=[DWDObservationPeriod.HISTORICAL],
-        start_date="1971-01-01",
-    )
-
-    assert request == DWDObservationData(
-        station_ids=[1],
-        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
-        resolution=DWDObservationResolution.DAILY,
-        periods=[
-            DWDObservationPeriod.HISTORICAL,
-        ],
-        start_date=pd.Timestamp("1971-01-01"),
-        end_date=pd.Timestamp("1971-01-01"),
-    )
-
     with pytest.raises(StartDateEndDateError):
         DWDObservationData(
             station_ids=[1],
@@ -183,6 +165,72 @@ def test_dwd_observation_data_time_input():
             start_date="1971-01-01",
             end_date="1951-01-01",
         )
+
+
+def test_dwd_observation_data_dynamic_period():
+    # Historical period expected
+    request = DWDObservationData(
+        station_ids=[1],
+        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
+        resolution=DWDObservationResolution.DAILY,
+        start_date="1971-01-01",
+    )
+
+    assert request.periods == [
+        DWDObservationPeriod.HISTORICAL,
+    ]
+
+    # Historical and recent period expected
+    request = DWDObservationData(
+        station_ids=[1],
+        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
+        resolution=DWDObservationResolution.DAILY,
+        start_date="1971-01-01",
+        end_date=pd.Timestamp(datetime.now()) - pd.Timedelta(days=400),
+    )
+
+    assert request.periods == [
+        DWDObservationPeriod.HISTORICAL,
+        DWDObservationPeriod.RECENT,
+    ]
+
+    # Historical, recent and now period expected
+    request = DWDObservationData(
+        station_ids=[1],
+        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
+        resolution=DWDObservationResolution.DAILY,
+        start_date="1971-01-01",
+        end_date=pd.Timestamp(datetime.now()),
+    )
+
+    assert request.periods == [
+        DWDObservationPeriod.HISTORICAL,
+        DWDObservationPeriod.RECENT,
+        DWDObservationPeriod.NOW,
+    ]
+
+    # !!!Recent and now period cant be tested dynamically
+    # TODO: add test with mocked datetime here
+
+    # Now period
+    request = DWDObservationData(
+        station_ids=[1],
+        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
+        resolution=DWDObservationResolution.DAILY,
+        start_date=pd.Timestamp(datetime.now()),
+    )
+
+    assert request.periods == [DWDObservationPeriod.NOW]
+
+    # No period (for example in future)
+    request = DWDObservationData(
+        station_ids=[1],
+        parameters=[DWDObservationParameterSet.CLIMATE_SUMMARY],
+        resolution=DWDObservationResolution.DAILY,
+        start_date=pd.Timestamp(datetime.now()) + pd.Timedelta(days=720),
+    )
+
+    assert request.periods == []
 
 
 def test_observation_data_storing():